Example of an API Key DelegatingHandler
In this example we also consider the api_key in the QueryString, which is useful when testing with Swagger.
However we don't actually check for the api_key existence or not here at the DelegationHandler, because this check
is a concern we should have at the AuthenticaionFilder stage of the pipeline.